Move component files out of glasslibrary package
