+++ src/org/labkey/test/components/glassLibrary/navigation/ProductMenu.java
@@ -5,8 +5,8 @@
 package org.labkey.test.components.glassLibrary.navigation;
 
 import org.labkey.test.Locator;
-import org.labkey.test.components.glassLibrary.components.MultiMenu;
 import org.labkey.test.components.html.BaseBootstrapMenu;
+import org.labkey.test.components.react.MultiMenu;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
