+++ src/org/labkey/test/components/react/OmniBox.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2019 LabKey Corporation. All rights reserved. No portion of this work may be reproduced in
  * any form or by any electronic or mechanical means without written permission from LabKey Corporation.
  */
-package org.labkey.test.components.glassLibrary.components;
+package org.labkey.test.components.react;
 
 import org.jetbrains.annotations.Nullable;
 import org.labkey.test.Locator;
