+++ src/org/labkey/test/components/react/files/FileSelectTree.java
@@ -1,14 +1,14 @@
-package org.labkey.test.components.glassLibrary.files;
+package org.labkey.test.components.react.files;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.components.Component;
 import org.labkey.test.components.WebDriverComponent;
 import org.labkey.test.components.html.Checkbox;
+import org.openqa.selenium.By;
 import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.By;
 
 import java.util.List;
