+++ src/org/labkey/test/components/react/files/FileUploadPanel.java
@@ -1,4 +1,4 @@
-package org.labkey.test.components.glassLibrary.components;
+package org.labkey.test.components.react.files;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
