+++ src/org/labkey/test/components/react/grids/EditableGrid.java
@@ -1,4 +1,4 @@
-package org.labkey.test.components.glassLibrary.grids;
+package org.labkey.test.components.react.grids;
 
 import org.apache.commons.lang3.SystemUtils;
 import org.labkey.test.Locator;
