+++ src/org/labkey/test/components/react/grids/GridRow.java
@@ -1,4 +1,4 @@
-package org.labkey.test.components.glassLibrary.grids;
+package org.labkey.test.components.react.grids;
 
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
