+++ src/org/labkey/test/pages/study/DatasetDesignerPage.java
@@ -4,12 +4,12 @@ import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.WebTestHelper;
 import org.labkey.test.components.domain.DomainDesigner;
-import org.labkey.test.components.glassLibrary.components.FilteringReactSelect;
-import org.labkey.test.components.glassLibrary.components.ReactSelect;
 import org.labkey.test.components.html.Checkbox;
 import org.labkey.test.components.html.Input;
 import org.labkey.test.components.html.RadioButton;
 import org.labkey.test.components.html.ToggleButton;
+import org.labkey.test.components.react.select.FilteringReactSelect;
+import org.labkey.test.components.react.select.ReactSelect;
 import org.labkey.test.components.study.AdvancedDatasetSettingsDialog;
 import org.labkey.test.pages.DatasetPropertiesPage;
 import org.openqa.selenium.WebDriver;
